Refactored persistence scheme
